feat : 현재 에셋 조회 및 변경 API 추가
Former-commit-id: 8c764a5070a7898f96f3b2f430ed51ffd93988bd

+++ backend/src/main/java/com/thedebuggers/backend/domain/entity/User.java
@@ -45,9 +45,9 @@ public class User {
 
     private int coin;
 
-    private int avatar;
+    private long avatar;
 
-    private int room;
+    private long room;
 
     @NotNull
     @Enumerated(EnumType.STRING)
